requirements.txt
black==22.3.0
Django==4.0.4
django-extensions==3.1.5
factory-boy==3.2.1
flake8==4.0.1
ipython==8.3.0
mypy==0.950
psycopg2==2.9.3
pylint-django==2.5.3
tblib==1.7.0
werkzeug[watchdog]==2.0.3
